requirements.txt
django==2.0.5
django-ckeditor==5.4.0
pillow==5.1.0

